requirements.txt
chainlit
langchain
openai
python-dotenv
pymupdf
qdrant-client
gunicorn